requirements.txt
pyautogen>=0.2.0
groq>=0.3.0
agentops>=0.2.0
python-dotenv>=0.19.0